Run publish sub-scripts _before_ publishing
Using scripty to default to running all publish:* works, but only when
the publish script is invoked. Unfortunately, the publish lifecycle
script is invoked _after_ publishing to npm. We'd rather publish to
github and homebrew _before_ npm, so we need to use prepublish.

Of course, because npm history, we have to use prepublishOnly.