Prefer our npm-installed version of update-version-defs
Having node-build-update-defs as an npm dep is sufficient to make it
available via `nodenv update-version-defs` (within the context of an npm
script) since npm prepends the local node_modules/.bin directory to
PATH. And since nodenv will respect any nodenv-* executable in path as a
subcommand, this all works fine.

_However_, nodenv will _prefer_ a nodenv-* executable that is found
within its $(nodenv root)/plugins directory. Thus, if a user has
node-build-update-defs installed as a nodenv plugin, that version would
be used instead of the version installed as an npm-dep local to this
package. To avoid this, this change will invoke the
nodenv-update-version-defs executable directly and _not_ go through
nodenv's subcommand bin resolution.